backups/backend/requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
requests==2.25.1

# Django
# ------------------------------------------------------------------------------
Django==3.2.6
djangorestframework~=3.12.4
django-cors-headers==3.7.0  # https://github.com/adamchainz/django-cors-headers
psycopg2==2.8.6  # https://github.com/psycopg/psycopg2
gunicorn==20.1.0
drf-yasg==1.20.0
